refactorng: moving converters to a seperate file
